CB-15763 Freeipa: Optimisticlockexception in StackUpdater caused termination flow interrupt
In the freeipa service, in some cases there might be parallel flows. In that case updating the stack status might happen at the same time, causing OptimisticLockException or StaleObjectStateExceptions.
This commit applies a retry logic around reading, updating and saving the Stack. In this way all status will get written to the DB. Also, the stack is read from DB again just before the save method, reducing the possibility of parallel updates.